demo/hello-world-vulnerable/requirements.txt
# 의도적으로 오래된 버전들 (취약점 포함)
Flask==1.0.0           # CVE-2019-1010083 (Debug mode RCE)
requests==2.19.0       # CVE-2018-18074 (Credential disclosure)
PyYAML==3.13          # CVE-2017-18342 (Arbitrary code execution)
Jinja2==2.10          # CVE-2019-10906 (Sandbox escape)
Werkzeug==0.14        # CVE-2019-14806 (Debug console)
urllib3==1.24.1       # CVE-2019-11324 (Certificate verification)
cryptography==2.3     # CVE-2018-10903 (GCM nonce reuse)
Pillow==5.2.0         # CVE-2019-16865 (Buffer overflow)
SQLAlchemy==1.2.0     # CVE-2019-7164 (SQL injection)
lxml==4.2.1           # CVE-2018-19787 (XXE vulnerability)